+++ examples/example/react/src/router.tsx
@@ -12,6 +12,7 @@ const LinkDemo = lazy(() => import('./pages/LinkDemo'))
 const TextDemo = lazy(() => import('./pages/TextDemo'))
 const CodeDemo = lazy(() => import('./pages/CodeDemo'))
 const KbdDemo = lazy(() => import('./pages/KbdDemo'))
+const HighlightDemo = lazy(() => import('./pages/HighlightDemo'))
 const ImageDemo = lazy(() => import('./pages/ImageDemo'))
 const ImageCropperDemo = lazy(() => import('./pages/ImageCropperDemo'))
 const ImageViewerDemo = lazy(() => import('./pages/ImageViewerDemo'))
@@ -166,6 +167,7 @@ const router = createHashRouter([
       { path: 'text', element: <TextDemo /> },
       { path: 'code', element: <CodeDemo /> },
       { path: 'kbd', element: <KbdDemo /> },
+      { path: 'highlight', element: <HighlightDemo /> },
       { path: 'image', element: <ImageDemo /> },
       { path: 'image-cropper', element: <ImageCropperDemo /> },
       { path: 'image-viewer', element: <ImageViewerDemo /> },