+++ examples/example/vue3/src/router.ts
@@ -22,6 +22,7 @@ const router = createRouter({
         { path: 'text', component: () => import('./pages/TextDemo.vue') },
         { path: 'code', component: () => import('./pages/CodeDemo.vue') },
         { path: 'kbd', component: () => import('./pages/KbdDemo.vue') },
+        { path: 'highlight', component: () => import('./pages/HighlightDemo.vue') },
         { path: 'image', component: () => import('./pages/ImageDemo.vue') },
         {
           path: 'image-cropper',